requirements.txt
absl-py                 2.2.2
annotated-types         0.7.0
asttokens               3.0.0
attrs                   25.3.0
beartype                0.20.2
blis                    1.3.0
catalogue               2.0.10
certifi                 2025.1.31
cffi                    1.17.1
charset-normalizer      3.4.1
click                   8.1.8
cloudpathlib            0.21.0
cloudpickle             3.1.1
colorama                0.4.6
comm                    0.2.2
confection              0.1.5
contourpy               1.3.1
cycler                  0.12.1
cymem                   2.0.11
debugpy                 1.8.14
decorator               5.2.1
executing               2.2.0
fastai                  2.8.1
fastcore                1.8.1
fastdownload            0.0.7
fastprogress            1.0.3
fasttransform           0.0.2
filelock                3.18.0
flatbuffers             25.2.10
fonttools               4.57.0
fsspec                  2025.3.2
idna                    3.10
ipykernel               6.29.5
ipython                 9.1.0
ipython_pygments_lexers 1.1.1
jax                     0.5.3
jaxlib                  0.5.3
jedi                    0.19.2
Jinja2                  3.1.6
joblib                  1.4.2
jupyter_client          8.6.3
jupyter_core            5.7.2
kiwisolver              1.4.8
langcodes               3.5.0
language_data           1.3.0
marisa-trie             1.2.1
markdown-it-py          3.0.0
MarkupSafe              3.0.2
matplotlib              3.10.1
matplotlib-inline       0.1.7
mdurl                   0.1.2
mediapipe               0.10.21
ml_dtypes               0.5.1
mpmath                  1.3.0
murmurhash              1.0.12
nest-asyncio            1.6.0
networkx                3.4.2
numpy                   2.2.5
opencv-contrib-python   4.11.0.86
opencv-python           4.11.0.86
opt_einsum              3.4.0
packaging               24.2
pandas                  2.2.3
parso                   0.8.4
pillow                  11.2.1
pip                     24.0
platformdirs            4.3.7
plum-dispatch           2.5.7
preshed                 3.0.9
prompt_toolkit          3.0.51
protobuf                4.25.6
psutil                  7.0.0
pure_eval               0.2.3
pycparser               2.22
pydantic                2.11.3
pydantic_core           2.33.1
Pygments                2.19.1
pyparsing               3.2.3
python-dateutil         2.9.0.post0
pytz                    2025.2
pywin32                 310
PyYAML                  6.0.2
pyzmq                   26.4.0
requests                2.32.3
rich                    14.0.0
scikit-learn            1.6.1
scipy                   1.15.2
sentencepiece           0.2.0
setuptools              65.5.0
shellingham             1.5.4
six                     1.17.0
smart-open              7.1.0
sounddevice             0.5.1
spacy                   3.8.5
spacy-legacy            3.0.12
spacy-loggers           1.0.5
srsly                   2.5.1
stack-data              0.6.3
sympy                   1.13.1
thinc                   8.3.6
threadpoolctl           3.6.0
torch                   2.6.0
torchvision             0.21.0
tornado                 6.4.2
tqdm                    4.67.1
traitlets               5.14.3
typer                   0.15.2
typing_extensions       4.13.2
typing-inspection       0.4.0
tzdata                  2025.2
urllib3                 2.4.0
wasabi                  1.1.3
wcwidth                 0.2.13
weasel                  0.4.1
wrapt                   1.17.2